pyproject.toml
[project]
name = "aaron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datetime>=6.0",
    "playsound3>=3.3.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-decouple>=3.8",
    "ruff>=0.15.4",
]
